+++ src/components/AboutMe.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+const AboutMe = (props) => {
+  return (
+    <>
+      <h2>About Me</h2>
+      <p>Software Engineer trying to overcome impostor syndrome.
+      I am married with a beautiful daughter & I have six cats.
+      Take that crazy cat lady stereotype!
+      </p>
+      <a className="link" href="mailto:giannelli.tech@gmail.com">Email me.</a>
+    </>
+  )
+}
+
+export default AboutMe