+++ src/components/StockQuotes/Dashboard.js
@@ -4,8 +4,7 @@ import { Link } from 'gatsby'
 const Dashboard = ({ children, uri }) => {
   return (
     <div>
-      <h2>Data Dashboard</h2>
-      <h3>Client-Only Route</h3>
+      <h2>Stock Quote Dashboard</h2>
       <nav>
         <Link to={`${uri}/tsla`}>Data for Tesla</Link>
         <Link to={`${uri}/aapl`}>Data for Apple</Link>