+++ src/components/StockQuotes/DashboardDefault.js
@@ -3,8 +3,7 @@ import React from 'react'
 const DashboardDefault = () => {
   return (
     <div>
-      <h2>Nested Dashboard Index</h2>
-      <p>This gets rendered when no data parameters are provided.</p>
+      <p>Select link above to view Stock Quote</p>
     </div>
   )
 }