+++ src/pages/about-me.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Layout from '../components/Layout';
+import AboutMe from '../components/AboutMe';
+
+const AboutMePage = (props) => {
+  return (
+    <Layout>
+      <AboutMe />
+    </Layout>
+  )
+}
+
+export default AboutMePage