+++ src/pages/app.js
@@ -1,29 +1,30 @@
 import React from 'react'
 import Layout from '../components/Layout'
 import { Router, Link } from "@reach/router"
-import MyInfo from '../components/MyInfo'
 import Default from '../components/Default'
-import Dashboard from '../components/Dashboard'
-import DashboardDefault from '../components/DashboardDefault'
-import Data from '../components/Data'
+import Dashboard from '../components/StockQuotes/Dashboard'
+import DashboardDefault from '../components/StockQuotes/DashboardDefault'
+import Data from '../components/StockQuotes/Data'
+import ChuckNorrisJokeGenerator from '../components/ChuckNorrisJokeGenerator'
 
 const App = ({ location }) => {
   return (
     <Layout>
-      <h1>Welcome to the App Page <span>{location.pathname}</span></h1>
+      <h1>Welcome to my Projects Page!</h1>
       <nav>
-        <Link to='/app/info'>Info</Link>
-        <Link to='/app/dashboard'>Dashboard</Link>
+          <Link to='/app'>Projects List</Link>
+        <Link to='/app/chuck-norris'>Chuck Norris Joke Generator</Link>
+        <Link to='/app/stock-quotes'>Stock Quotes Search</Link>
       </nav>
       <Router basepath="/app">
         <Default path="/" />
-        <MyInfo path="/info" />
-        <Dashboard path="/dashboard">
+        <ChuckNorrisJokeGenerator path="/chuck-norris" />
+        <Dashboard path="/stock-quotes">
           <DashboardDefault path="/" />
           <Data path=":dataId" />
         </Dashboard>
       </Router>
-    </Layout>
+    </Layout >
   )
 }