+++ src/pages/index.js
@@ -4,8 +4,7 @@ import Layout from '../components/Layout'
 const Home = ({ location }) => {
   return (
     <Layout>
-      <h1>Home Page <span>{location.pathname}</span></h1>
-      <h2>Static HTML Page</h2>
+      <h1>Welcome!</h1>
     </Layout>
   )
 }