+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-01-02
+- Update Homepage message
+- Update App Nav link to "Projects"
+- Update Projects page 
+- Update ChuckNorrisJokeGenerator to be project component instead of page
+- Updated views for Stock Quote Generator
+- Add initial About Me page
+
 ## [0.1.0] - 2021-01-02
 - Created MPA using Gatsby with API for stock data https://rapidapi.com/blog/react-multi-page-app/
 - Added in the Chuck Norris generator (Okta auth to come later ) https://developer.okta.com/blog/2019/03/06/simple-user-authentication-in-react